==> 2_Feature_building.ipynb <==
# Building features from the processed data after initial exploration

# Importing the required libraries
import numpy as np
import os
import pandas as pd
from ast import literal_eval
# Chaning back to the resource directory if it is not there
if("Res" not in os.getcwd()):
    os.chdir("..")

# Reading the initial data processed from the code initial EDA
intial_df = pd.read_csv("Res/processed_data/final_spam_ham_df.csv",sep=",",
                        converters={"tokens": literal_eval,"useful_tokens": literal_eval,
                                   "stemmed_tokens": literal_eval,"lemma_tokens": literal_eval})
intial_df.head()

all_useful_tokens = []
# Using the lemma tokens as the useful tokens
for useful_token in intial_df["lemma_tokens"]:
    all_useful_tokens.extend(useful_token)
all_useful_tokens = list(map(lambda x: x.lower(), all_useful_tokens))

from collections import Counter
all_useful_tokens_dict = dict(Counter(all_useful_tokens))

# Finding the top 3000 tokens based on frequency of occurance
useful_tokens_dict = {}
for key,value in all_useful_tokens_dict.iteritems():
    if(value>220 and len(key)>2):
        useful_tokens_dict[key.lower()] = value

with open('Res/useful_tokens_dict.txt','w') as data:
    data.write(str(useful_tokens_dict))
    
def intersect_useful_tokens(lemma_tokens_row):
    return (list(set(map(lambda x: x.lower(), lemma_tokens_row))&set(useful_tokens_dict.keys())))
intial_df["attributes"] = intial_df.apply(lambda row: intersect_useful_tokens(row["lemma_tokens"]),axis=1)
intial_df["attributes_len"] = intial_df.apply(lambda row: float(len(row["attributes"])),axis=1)
intial_df["corpus"] = intial_df.apply(lambda row: " ".join(row["attributes"]),axis=1)

# Building the tf-idf data frame
from sklearn.feature_extraction.text import TfidfVectorizer
corpus = list(intial_df["corpus"])
vectorizer = TfidfVectorizer()
X = vectorizer.fit_transform(corpus)
attributes_df = pd.DataFrame(X.todense(),columns = vectorizer.get_feature_names())
intial_df = pd.concat([intial_df,attributes_df],axis=1)

# Writing the final features data into attributes_df.csv file
attributes_hot_encode = intial_df
attributes_hot_encode.to_csv("Res/processed_data/attributes_df.csv",sep=",",index=False)
with open("Res/processed_data/attributes_names.txt", 'w') as f:
    for item in vectorizer.get_feature_names():
        f.write("%s\n" % item)

print("Completed, Number of attributes created = {}".format(len(vectorizer.get_feature_names())))
