==> 5_Clean_raw_data.ipynb <==
import os
import pandas as pd

pd.set_option('display.max_columns', 500)
pd.set_option('display.width', 1000)

os.chdir("..")

dir_list = os.listdir("Res/raw_data/")
dir_list = filter(lambda x: "." not in x, dir_list)

df = pd.DataFrame(columns=["owner","filepath","mail_label","content_mail"])

count=0
for dir_1 in dir_list:
    accepted_mail = []
    mail_label = "spam" if(dir_1 in ("SH","GP","BG")) else "ham"
    print(dir_1),
    for dir_2 in (os.listdir("Res/raw_data/"+dir_1)):
        current_dir=[]
        if("." not in dir_2):
            current_dir = os.listdir("Res/raw_data/"+dir_1+"/"+dir_2)
            try:
                for x in current_dir:
                    current_dir = ["Res/raw_data/" + dir_1 + "/" + dir_2 +"/"+x+"/"+y
                                   for y in os.listdir("Res/raw_data/" + dir_1 + "/" + dir_2 + "/" + x)]

            except:
                current_dir = ["Res/raw_data/" + dir_1 + "/" + dir_2 +"/"+x
                               for x in os.listdir("Res/raw_data/" + dir_1 + "/" + dir_2)]

        for file_name in current_dir:
            file = open(file_name,"r")
            file_content = file.read()
            if("content-type: text/plain" in file_content.lower() or "content-type: text/html" in file_content.lower()):
                df.loc[count] = [dir_1,file_name,mail_label,file_content]
                count+=1
                accepted_mail.append(file_name)
    print(len(accepted_mail))

df.to_csv("Res/accepted_raw_data/filtered_raw_data.csv",sep=",",index=False)
