==> 7_Study_raw_data.ipynb <==
# Studying the effects of changing the data and running the same models and creating new models

# Importing the required libraries
import pandas as pd
import numpy as np
import os
os.chdir("..")

from sklearn.cross_validation import train_test_split

enron7 = pd.read_csv("Res/Complete_data_processed/enron7.csv",sep=",")
enron8 = pd.read_csv("Res/Complete_data_processed/enron8.csv",sep=",")
enron9 = pd.read_csv("Res/Complete_data_processed/enron9.csv",sep=",")
enron10 = pd.read_csv("Res/Complete_data_processed/enron10.csv",sep=",")
enron11 = pd.read_csv("Res/Complete_data_processed/enron11.csv",sep=",")
enron12 = pd.read_csv("Res/Complete_data_processed/enron12.csv",sep=",")

# Studying the accuracy using the xgboost models built from enron 1 to enron 6 dataset
new_enron_accuracy_df = pd.DataFrame(columns=["Enron_dataset","model_name","spam","ham","tp","tn","fp","fn"
                                              ,"acc","recall","specificity"])
enron_dataset_list_df = [enron7,enron8,enron9,enron10,enron11,enron12]
new_enron_list = ["Enron"+str(i) for i in range(7,13)]
i=0
j=-1
for enron_dataset in enron_dataset_list_df:
    j+=1
    print("Working on the "+new_enron_list[j])
    model_no_list = range(1,7)
    # Adding some of the attributes that are missing in the data that we created now to test against the models 1 to 6
    for col in ['additiona','aiso','andmanyother','beiieves','couid','dbcaps','enronxgate','hplno',
                        'inciude','inciuding','invoive','lst','maiiings','majordomo','newsietter',
                        'piaced','pians','potentia','projecthoneypot','pubiisher','resuits','shouid',
                        'squirrelmail','technoiogies','technoiogy','utf']:
        enron_dataset[col]=0.0
    for model_no in model_no_list:
        print("Running the model trained from Enron data : "+str(model_no))
        col_attributes = []
        with open("Res/processed_data/attributes_names.txt", 'r') as f:
            col_attributes.extend(f.read().split())
        baseline_X_features = ['line_count', 'token_count', 'punctuations_count', 'single_char_count', 
                                 'number_token_count', 'year_count', 'stopword_count','median_useful_token_len', 
                                 'avg_useful_token_len', 'rareword_count','attributes_len']
        baseline_X_features.extend(col_attributes)

        import pickle
        import xgboost as xgb
        # load model from file
        loaded_xgb = pickle.load(open("Res/"+"enron"+str(model_no)+"_xgboost.pickle.dat", "rb"))
        
        spam=enron_dataset.loc[enron_dataset["mail_label"]=="spam"]
        ham=enron_dataset.loc[enron_dataset["mail_label"]=="ham"]
        
        enron_dataset["label"] = pd.Series(enron_dataset["mail_label"]).apply(lambda x: 1.0 if x=="spam" else 0.0)
        dtest = xgb.DMatrix(enron_dataset[baseline_X_features], label=enron_dataset["label"])
        y_pred = loaded_xgb.predict(dtest)
        y_probas = y_pred # predicted probabilities generated by sklearn classifier
        y_pred = pd.Series(y_pred).apply(lambda x: 1.0 if x>0.5 else 0.0)

        dfs = pd.DataFrame(columns = ["Actual","Pred"])
        dfs["Actual"]=enron_dataset["label"]
        dfs["Pred_probs"]=loaded_xgb.predict(dtest)
        dfs["Pred"] = dfs.apply(lambda row:1.0 if row["Pred_probs"]>0.5 else 0.0,axis=1)

        from sklearn.metrics import confusion_matrix

        y_test = dfs["Actual"]
        y_pred = dfs["Pred"]

        tn, fp, fn, tp = confusion_matrix(y_test, y_pred).ravel()
        print("True Positives(spam) = {}".format(tp))
        print("True Negatives(ham)  = {}".format(tn))
        print("False Positives(actual=ham, predicted=spam) = {}".format(fp))
        print("False Negatives(actual=spam,predicted=ham) = {}".format(fn))
        print("Accuracy = {:.2f}%".format((tp+tn)/float(len(y_test))*100))
        print("Recall = {:.2f}%".format((tp/float(tp+fn))*100))
        print("Specificity = {:.2f}%".format((tn/float(tn+fp))*100))
        
        new_enron_accuracy_df.loc[i]=[new_enron_list[j],"enron"+str(model_no)+"_xgboost",
                                      len(spam),len(ham),tp, tn, fp, fn,
                                      (tp+tn)/float(len(spam)+len(ham))*100,
                                      (tp/float(tp+fn))*100,
                                      (tn/float(tn+fp))*100]
        i+=1
        print("\n")
new_enron_accuracy_df.to_csv("Res/Complete_data_processed/new_enron_accuracy.csv",sep=",",index=False)

# Creating indiviual xgboost models for enron 7 to 12 dataset separately
enron_measure_model_df = pd.DataFrame(columns=["Enron_dataset","spam","ham","train_spam","train_ham",
                                            "test_spam","test_ham","tp","tn","fp","fn","acc","recall","specificity"])

enron_dataset_list_df = [enron7,enron8,enron9,enron10,enron11,enron12]
j=0
for ernron_df in enron_dataset_list_df:
    
    col_attributes=[]
    with open("Res/processed_data/attributes_names.txt", 'r') as f:
            col_attributes.extend(f.read().split())
    # Baseline Features for the logistic regression model
    baseline_X_features = ['line_count', 'token_count', 'punctuations_count', 'single_char_count', 
                         'number_token_count', 'year_count', 'stopword_count','median_useful_token_len', 
                         'avg_useful_token_len', 'rareword_count','attributes_len']
    baseline_X_features.extend(col_attributes)
    Y_feature = "mail_label"

    # Subsetting the data as X and y
    spam = ernron_df[ernron_df["mail_label"]=="spam"]
    ham = ernron_df[ernron_df["mail_label"]=="ham"]

    X_train_spam, X_test_spam, y_train_spam, y_test_spam = \
    train_test_split(spam[baseline_X_features], spam[Y_feature], test_size=0.2, random_state=0)

    X_train_ham, X_test_ham, y_train_ham, y_test_ham = \
    train_test_split(ham[baseline_X_features], ham[Y_feature], test_size=0.2, random_state=0)

    X_train = pd.concat([X_train_spam,X_train_ham])
    X_test = pd.concat([X_test_spam,X_test_ham])
    y_train = pd.concat([y_train_spam,y_train_ham])
    y_test = pd.concat([y_test_spam,y_test_ham])
    y_train["label"] = y_train.apply(lambda x: 1.0 if x=="spam" else 0.0)
    y_test["label"] = y_test.apply(lambda x: 1.0 if x=="spam" else 0.0)

    dtrain = xgb.DMatrix(X_train, label=y_train["label"])
    dtest = xgb.DMatrix(X_test, label=y_test["label"])

    param = {'max_depth': 100, 'eta': 1, 'verbosity': 0, 'objective': 'binary:logistic',"n_estimators": [200]}
    param['nthread'] = 4
    param['eval_metric'] =  ['auc']

    evallist = [(dtest, 'eval'), (dtrain, 'train')]

    num_round = 30
    clf = xgb.train(param, dtrain, num_round, evallist)

    y_pred = clf.predict(dtest)
    y_probas = y_pred # predicted probabilities generated by sklearn classifier
    y_pred = pd.Series(y_pred).apply(lambda x: 1.0 if x>0.5 else 0.0)

    tn, fp, fn, tp = confusion_matrix(y_test["label"], y_pred).ravel()
    print("True Positives(spam) = {}".format(tp))
    print("True Negatives(ham)  = {}".format(tn))
    print("False Positives(actual=ham, predicted=spam) = {}".format(fp))
    print("False Negatives(actual=spam,predicted=ham) = {}".format(fn))
    print("Accuracy = {:.2f}%".format((tp+tn)/float(len(y_test))*100))
    print("Recall = {:.2f}%".format((tp/float(tp+fn))*100))
    print("Specificity = {:.2f}%".format((tn/float(tn+fp))*100))
    
    enron_measure_model_df.loc[j]= ["Enron"+str(j+7),
           len(spam),len(ham),len(X_train_spam),len(X_train_ham),
           len(X_test_spam),len(X_test_ham),
           tp,tn,fp,fn,
           (tp+tn)/float(len(y_test))*100,
           (tp/float(tp+fn))*100,
           (tn/float(tn+fp))*100]
    j+=1
enron_measure_model_df.to_csv("Res/Complete_data_processed/enron_measure_model.csv",sep=",",index=False)

enron_measure_model_df

new_enron_accuracy_df

print("Completed!")
